backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Celery + Redis
celery==5.4.0
redis==5.1.1

# File conversion - Images
Pillow==10.4.0
Wand==0.6.13

# File conversion - Audio/Video
ffmpeg-python==0.2.0

# File conversion - Documents
python-pptx==1.0.2
python-docx==1.1.2
openpyxl==3.1.5
pypdf==4.3.1
markdown==3.7
weasyprint==62.3

# File conversion - Data formats
PyYAML==6.0.2
toml==0.10.2
xmltodict==0.13.0
dicttoxml==1.7.16

# File conversion - Ebooks (uses calibre CLI externally)
# File conversion - Fonts
fonttools==4.54.1
brotli==1.1.0
zopfli==0.2.3

# File conversion - Archives
py7zr==0.22.0

# Utilities
python-magic-bin==0.4.14
aiofiles==24.1.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
